ebib-get-field-value: support inheritance from @XData entries
This means that more one than entry might be inherited from at one
time. Accordingly, store entries to inherit from as a
list (`xref-key-alist'). xdata and crossref inheritances have
different behaviour and precedence, so this needs to be an alist
including the type of the entry inherited from. The list is
constructed such that earlier entries take precedence in inheritance.

Whenever a value is inherited, retain the key it comes from as a
string in `xref-inherit-from-key'. Use this variable to add
appropriate text properties to the returned result.