Correctly propertize all cross referenced inheritances
ebib-get-field-value: propertize inherited xref-value when inheriting
a whole field. Remove code which assumes the whole field is always
what is inherited (granular xdata references mean it might not be).

ebib--get-field-highlighted: Be more specific about where the crossref
face is applied (only to the parts of a string with an `ebib--xref'
property).